requirements.txt
fastapi==0.114.0
uvicorn==0.30.6
python-dotenv==1.0.1
sqlalchemy==2.0.34
psycopg2-binary==2.9.9
PyJWT ==2.9.0
passlib[bcrypt]==1.7.4
pydantic_settings==2.5.0
python-multipart==0.0.9
email_validator==2.2.0